src/logger.ts => src/logger/logger.ts, src/logger/factory.ts
