+++ src/logger/factory.ts
@@ -1,17 +1,10 @@
 import * as Pino from 'pino';
 const pino = require('pino');
 
-export type ILogger = Pino.Logger;
-
-export type Level = Pino.Level;
-
-export interface ILoggerOptions {
-    level?: Level;
-    machineReadable?: boolean;
-}
+import { ILogger, ILoggerOptions } from './logger';
 
 // Create a logger.
-export function createLogger(options: ILoggerOptions = {}) {
+export function createLogger(options: ILoggerOptions = {}): ILogger {
     return pino({
         level: options.level ? options.level : 'error',
         prettyPrint: 'machineReadable' in options ? !options.machineReadable : true
