+++ src/logger/index.ts
@@ -1,0 +1,2 @@
+export * from './factory';
+export * from './logger';
